back to uint64_t

+++ valhalla/mjolnir/osmrestriction.h
@@ -70,22 +70,22 @@ struct OSMRestriction {
   /**
    * Set the from way id
    */
-  void set_from(uint32_t from);
+  void set_from(uint64_t from);
 
   /**
    * Get the from way id
    */
-  uint32_t from() const;
+  uint64_t from() const;
 
   /**
    * Set the to way id
    */
-  void set_to(uint32_t to);
+  void set_to(uint64_t to);
 
   /**
    * Get the to way id
    */
-  uint32_t to() const;
+  uint64_t to() const;
 
   /**
    * Set the time domain
@@ -129,10 +129,10 @@ struct OSMRestriction {
   }
 
   // from is a way - uses OSM way Id.
-  uint32_t from_;
+  uint64_t from_;
 
   // to is a way - uses OSM way Id.
-  uint32_t to_;
+  uint64_t to_;
 
   // Via is a node. When parsing OSM this is stored as an OSM node Id.
   // It later gets changed into a GraphId.
